docs: add TODO section
I was considering using logical properties to support multilingual
styling in css-tensei but shorthand logical properties aren't really a
thing right now. It may be fine to write `padding-inline` and
`padding-block` instead of just `padding` but it's not fine to have to
write four different values for `border-radius`.

Either we use logical properties completely or we don't.